# Code-Companion/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  handle: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  town: string;
  university: string;
  /** @nullable */
  avatarUrl?: string | null;
  followersCount?: number;
  followingCount?: number;
  postsCount?: number;
  weixBucks?: number;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  handle: string;
  /** @minLength 1 */
  displayName: string;
  bio?: string;
  town: string;
  university: string;
  avatarUrl?: string;
}

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Post {
  id: number;
  authorHandle: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  authorUniversity?: string;
  content: string;
  townTag: string;
  /** @nullable */
  mediaUrl?: string | null;
  likesCount: number;
  repliesCount: number;
  repostsCount: number;
  liked?: boolean;
  createdAt: string;
}

export interface PostInput {
  authorHandle: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  townTag: string;
  mediaUrl?: string;
}

export interface Reply {
  id: number;
  postId: number;
  authorHandle: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface ReplyInput {
  authorHandle: string;
  /** @minLength 1 */
  content: string;
}

export type RelayStatus = typeof RelayStatus[keyof typeof RelayStatus];


export const RelayStatus = {
  online: 'online',
  degraded: 'degraded',
  offline: 'offline',
} as const;

export interface Relay {
  id: number;
  name: string;
  town: string;
  university: string;
  tier: number;
  status: RelayStatus;
  eventsPerHour: number;
  connectedPeers: number;
  uptimePercent: number;
  storageUsedMb?: number;
  region?: string;
  createdAt: string;
}

export interface NetworkStats {
  totalRelays: number;
  onlineRelays: number;
  totalUsers: number;
  totalPosts: number;
  activeTowns: number;
  weixBucksInCirculation: number;
  eventsLast24h: number;
  meshConnections?: number;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  post: 'post',
  join: 'join',
  relay_up: 'relay_up',
  relay_down: 'relay_down',
  milestone: 'milestone',
  trending: 'trending',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  description: string;
  town: string;
  /** @nullable */
  userHandle?: string | null;
  createdAt: string;
}

export interface ArchitectureLayer {
  layer: string;
  standardRole: string;
  blkspaceImpl: string;
  rationale: string;
  color?: string;
}

export type ArchitectureDataPrinciplesItem = {
  title: string;
  description: string;
};

export type ArchitectureDataSecurityLayersItem = {
  layer: string;
  mechanism: string;
  protection: string;
};

export interface ArchitectureData {
  layers: ArchitectureLayer[];
  principles: ArchitectureDataPrinciplesItem[];
  securityLayers: ArchitectureDataSecurityLayersItem[];
}

export type ListPostsParams = {
town?: string;
trending?: boolean;
};