# Code-Companion/lib/api-zod/src/generated/types/architectureData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ArchitectureDataPrinciplesItem } from './architectureDataPrinciplesItem';
import type { ArchitectureDataSecurityLayersItem } from './architectureDataSecurityLayersItem';
import type { ArchitectureLayer } from './architectureLayer';

export interface ArchitectureData {
  layers: ArchitectureLayer[];
  principles: ArchitectureDataPrinciplesItem[];
  securityLayers: ArchitectureDataSecurityLayersItem[];
}