# Code-Companion/lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  authorHandle: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  townTag: string;
  mediaUrl?: string;
}