# Code-Companion/lib/api-zod/src/generated/types/relay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RelayStatus } from './relayStatus';

export interface Relay {
  id: number;
  name: string;
  town: string;
  university: string;
  tier: number;
  status: RelayStatus;
  eventsPerHour: number;
  connectedPeers: number;
  uptimePercent: number;
  storageUsedMb?: number;
  region?: string;
  createdAt: string;
}