# Code-Companion/lib/api-zod/src/generated/types/reply.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Reply {
  id: number;
  postId: number;
  authorHandle: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}