# Code-Companion/lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  handle: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  town: string;
  university: string;
  /** @nullable */
  avatarUrl?: string | null;
  followersCount?: number;
  followingCount?: number;
  postsCount?: number;
  weixBucks?: number;
  createdAt: string;
}