# Code-Companion/lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 1 */
  handle: string;
  /** @minLength 1 */
  displayName: string;
  bio?: string;
  town: string;
  university: string;
  avatarUrl?: string;
}