# Code-Companion/lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlkSpace federated social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}